+++ src/app/Sidebar/components/searchInput.ts
@@ -17,7 +17,7 @@ const searchInput = (
       <div class="codicon codicon-search" aria-label="search"></div>
     </div>
   </div>
-  <p id="documentation-found-length" class="italic text-slate-400">${subtitle}</p>`;
+  <p id="documentation-found-length" class="italic">${subtitle}</p>`;
 };
 
 export default searchInput;