+++ src/lib/utils/documentation/getDocumentationViewContent.ts
@@ -37,15 +37,25 @@ const getDocumentationViewContent = (
     webview,
     extensionUri
   )}">
-  </script>
-  <script nonce="${nonce}" src="https://unpkg.com/@dotlottie/player-component@latest/dist/dotlottie-player.mjs" type="module"></script> 
+      </script>
+      <script nonce="${nonce}" src="https://unpkg.com/@dotlottie/player-component@latest/dist/dotlottie-player.mjs" type="module"></script> 
+      <script nonce="${nonce}" type="module" src="https://cdn.jsdelivr.net/npm/ldrs/dist/auto/zoomies.js"></script>
       </head>
       <body>
         <div id="no-documentation-found" class="flex flex-col gap-4 px-6 py-4">
-          <p>No documentation found. Try to reload the extension.</p>
-          <div id="reload" class="flex items-center justify-center gap-2 rounded bg-sky-500 p-2 text-slate-50   hover:cursor-pointer hover:bg-sky-400">
-          <div class="codicon codicon-refresh" aria-label="refresh"></div>
-          <p class="text-slate-50">Reload</p>
+          <p>No documentation found or loading too long ? Try to reload the extension.</p>
+          <div class="flex w-full flex-col items-center justify-center gap-6">
+            <div id="reload" class="flex w-full items-center justify-center gap-2 rounded bg-sky-500 p-2 text-slate-50 hover:cursor-pointer hover:bg-sky-400">
+              <div class="codicon codicon-refresh" aria-label="refresh"></div>
+              <p class="text-slate-50">Reload</p>
+            </div>
+            <l-zoomies
+              size="150"
+              stroke="4"
+              bg-opacity="0.1"
+              speed="1.4"
+              color="black" 
+            ></l-zoomies>
           </div>
         </div>
         <div id="documentation-container" class="overflow-hidden max-w-full"></div>