+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning].
 ### Changed
 
 - Get the documentation url, not the package home page
+- Use VSCode's css variables for text color instead of tailwind's gray color
+- Add a loader to the launch page
 
 ## [0.2.1] - 2024-07-17